requirements.txt
langchain
langchain-classic
langchain-groq
langchain-core
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
numpy
python-dotenv
requests
duckduckgo-search
streamlit
